+++ apps/extension/src/common/connect-event-emitter-with-port.ts
@@ -1,23 +1,31 @@
 import browser from 'webextension-polyfill'
 import { WildcardEventEmitter } from './wildcard-event-emitter'
 
-export const createEventEmitterFromPort = (port: browser.Runtime.Port) => {
-  const eventEmitter = new WildcardEventEmitter()
-
+export const connectEventEmitterWithPort = (
+  port: browser.Runtime.Port,
+  eventEmitter: WildcardEventEmitter
+) => {
   const portMessageListener = (message: any) => {
     if (message.type === 'emitEvent') {
-      eventEmitter.emit(message.data.event, ...message.data.args)
+      eventEmitter.emitQuitely(message.data.event, ...message.data.args)
     }
   }
 
   port.onMessage.addListener(portMessageListener)
 
+  const listener = (event: string, ...args: any[]) => {
+    port.postMessage({ type: 'emitEvent', data: { event, args } })
+  }
+
+  eventEmitter.on('*', listener)
+
   const handleDisconnect = () => {
+    eventEmitter.off('*', listener)
     port.onMessage.removeListener(portMessageListener)
     port.onDisconnect.removeListener(handleDisconnect)
   }
 
   port.onDisconnect.addListener(handleDisconnect)
 
-  return { eventEmitter, destroy: () => handleDisconnect() }
+  return { disconnect: () => handleDisconnect() }
 }
